+++ tests/ngtcp2_test_helper.h
@@ -61,7 +61,7 @@
  * This function returns the number of bytes written to |out|.
  */
 size_t ngtcp2_t_encode_stream_frame(uint8_t *out, uint8_t flags,
-                                    uint32_t stream_id, uint64_t offset,
+                                    uint64_t stream_id, uint64_t offset,
                                     uint16_t datalen);
 
 /*
